S57: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31605 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/s57/s57classregistrar.cpp
@@ -321,7 +321,7 @@ int S57ClassRegistrar::LoadInfo( const char * pszDirectory,
     if( fp != NULL )
         VSIFCloseL( fp );
 
-    nAttrCount = anAttrIndex.size();
+    nAttrCount = static_cast<int>(anAttrIndex.size());
 
 /* -------------------------------------------------------------------- */
 /*      Sort index by acronym.                                          */
